\name{Age}
\alias{Age}
\docType{data}
\title{Ability to Speak English by Age}
\description{
Cross-classisfication of people in Hawaii according to each group's age and ability to speak English. The table gives the number of people who speak English in Hawaii by age group. It excludes those less than 5 years old or speak only English.
}
\usage{data(Age)}
\format{
  A table of counts, with classifying factors Age (levels 1:5) and Ablilty (a person's ability to speak English; levels 1:3).
  \describe{
    \item{\code{Age}}{The various age groups include those 19 years or under, 20-39 years old, 40-59 years old, 60-79 years old, and 80 years and above}
    \item{\code{Well}}{Number of people who speak English well}
    \item{\code{Slightly}}{Number of people who don't speak English well}
    \item{\code{Never}}{Number of people who don't speak English at all}
  }
}

\source{
State of Hawaii Office of Language Access (http://hawaii.gov/dlir). Date released 2007. License: CC BY.
}

\keyword{datasets}
