\name{honoAP}
\alias{honoAP}
\docType{data}
\title{
Ambient Levels of Airborne Particulates (PM) in Honolulu
}
\description{
State of Hawaii Department Of Health Environmental Indicators.
}
\usage{data(honoAP)}
\format{
  A time series with 7 observations on the following 3 variables.
  \describe{
    \item{\code{year}}{Year}
    \item{\code{honoAvg}}{Honolulu Annual Average}
    \item{\code{national}}{National Standard}
  }
}

\source{
Hawaii State Department of Health (http://hawaii.gov/doh). Released July 31, 2012. License: CC BY.
}

\examples{
data(honoAP)
str(honoAP) 

}
\keyword{datasets}
