\name{libHoldings}
\alias{libHoldings}
\docType{data}
\title{
Library Holdings 1984-2011 
}
\description{
This data set contains information on Hawaii's State Public Library quantity of holdings for the years 1984 through 2011.
}
\usage{data(libHoldings)}
\format{
A data frame with 52 observations on the following 30 variables.
  \describe{
    \item{\code{island}}{Island where library is located}
    \item{\code{library}}{Town/ Census-Designated Place where library is located}
    \item{\code{FY84}}{Number of holdings in 1984}
    \item{\code{FY85}}{Number of holdings in 1985}
    \item{\code{FY86}}{Number of holdings in 1986}
    \item{\code{FY87}}{Number of holdings in 1987}
    \item{\code{FY88}}{Number of holdings in 1988}
    \item{\code{FY89}}{Number of holdings in 1989}
    \item{\code{FY90}}{Number of holdings in 1990}
    \item{\code{FY91}}{Number of holdings in 1991}
    \item{\code{FY92}}{Number of holdings in 1992}
    \item{\code{FY93}}{Number of holdings in 1993}
    \item{\code{FY94}}{Number of holdings in 1994}
    \item{\code{FY95}}{Number of holdings in 1995}
    \item{\code{FY96}}{Number of holdings in 1996}
    \item{\code{FY97}}{Number of holdings in 1997}
    \item{\code{FY98}}{Number of holdings in 1998}
    \item{\code{FY99}}{Number of holdings in 1999}
    \item{\code{FY00}}{Number of holdings in 2000}
    \item{\code{FY01}}{Number of holdings in 2001}
    \item{\code{FY02}}{Number of holdings in 2002}
    \item{\code{FY03}}{Number of holdings in 2003}
    \item{\code{FY04}}{Number of holdings in 2004}
    \item{\code{FY05}}{Number of holdings in 2005}
    \item{\code{FY06}}{Number of holdings in 2006}
    \item{\code{FY07}}{Number of holdings in 2007}
    \item{\code{FY08}}{Number of holdings in 2008}
    \item{\code{FY09}}{Number of holdings in 2009}
    \item{\code{FY10}}{Number of holdings in 2010}
    \item{\code{FY11}}{Number of holdings in 2011}
  }
}

\source{
These data are provided by the Hawaii State Public Library System (http://www.librarieshawaii.org/). Released June, 2012. License: Public Domain.
}

\examples{
data(libHoldings)
str(libHoldings) ;
}
\keyword{datasets}
