\name{libInternet}
\alias{libInternet}
\docType{data}
\title{
Library Internet Sessions By Year
}
\description{
Number of internet sessions in use for libraries throughout the State of Hawaii from 2006 to 2011.
}
\usage{data(libInternet)}
\format{
A multiple time series with 6 observations on the following 52 variables.
  \describe{		
    \item{\code{Year}}{Year}
    \item{\code{Hilo}}{Hilo Public Library}
    \item{\code{Honokaa}}{Honokaa Public Library}
    \item{\code{Kailua}}{Kailua-Kona Public Library}
    \item{\code{Keaau}}{Keaau Public Library}
    \item{\code{Keala}}{Kealakekua Public Library}
    \item{\code{Laupa}}{Laupahoehoe Public Library}
    \item{\code{View}}{Mountain View Public Library}
    \item{\code{Naale}}{Naalehu Public Library}
    \item{\code{Kohala}}{North Kohala Public Library}
    \item{\code{Pahala}}{Pahala Public Library}
    \item{\code{Pahoa}}{Pahoa Public Library}
    \item{\code{Parker}}{Thelma Parker Public Library}
    \item{\code{Hana}}{Hanapepe Public Library}
    \item{\code{Kapaa}}{Kapaa Public Library} 
    \item{\code{Koloa}}{Koloa Public Library}
    \item{\code{Lihue}}{Lihue Public Library}
    \item{\code{Prince}}{Princeville Public Library}
    \item{\code{Waimea}}{Waimea Public Library}
    \item{\code{Kahului}}{Kahului Public Library}
    \item{\code{Kihei}}{Kihei Public Library}
    \item{\code{Lahaina}}{Lahaina Public Library}
    \item{\code{Lanai}}{Lanai Public Library}
    \item{\code{Makawao}}{Makawao Public Library}
    \item{\code{Molokai}}{Molokai Public Library}
    \item{\code{Wailuku}}{Wailuku Public Library}
    \item{\code{Aiea}}{Aiea Public Library}
    \item{\code{Aina}}{Aina Haina Public Library}
    \item{\code{Ewa}}{Ewa Beach Public Library}
    \item{\code{Hawaii}}{Hawaii Kai Public Library}
    \item{\code{HSL}}{Hawaii State Library}
    \item{\code{Kahuku}}{Kahuku Public Library}
    \item{\code{Kailua}}{Kailua-Kona Public Library}
    \item{\code{Kaimuki}}{Kaimuki Public Library}
    \item{\code{Kalihi}}{Kalihi-Palama Public Library}
    \item{\code{Kaneohe}}{Kaneohe Public Library}
    \item{\code{Kapolei}}{Kapolei Public Library}
    \item{\code{LBPH}}{Library for the Blind and Physically Handicapped}
    \item{\code{Liliha}}{Liliha Public Library}
    \item{\code{Manoa}}{Manoa Public Library}
    \item{\code{McCully}}{McCully Public Library}
    \item{\code{Mililani}}{Mililani Public Library}
    \item{\code{Pearlc}}{Pearl City Public Library}
    \item{\code{Saltlake}}{Salt Lake Public Library}
    \item{\code{Wahiawa}}{Wahiawa Public Library}
    \item{\code{Waialua}}{Waialua Public Library}
    \item{\code{Waianae}}{Waianae Public Library}
    \item{\code{Waikiki}}{Waikiki Public Library}
    \item{\code{Waima}}{Waimanalo Public Library}
    \item{\code{Waipa}}{Waipahu Public Library}
    \item{\code{Bond}}{Bond Memorial Public Library}
    \item{\code{Holua}}{Holualoa}
    }
}
\source{
Hawaii State Public Library System (http://hawaii.sdp.sirsi.net/custom/web/). Released in 2011. License: CC BY.
}
\examples{
data(libInternet)
str(libInternet)
}
\keyword{datasets}